Actualizaciones en InterfazMainPrueba y InterfazVentana.
En InterfazMainPrueba se decidió eliminar todo porque es la ventana principal, no debería de tener lo que tendría la ventana emergente.
En nterfazVentana se agregó todo el contenido que antes tenía InterfazMainPrueba, esta si debería tener lo que es la ventana emergente ya que es la ventana emergente.
